feat: extract root component into separate file
